Move to new header and split PasswordLock
