+++ src/assets/scss/_variables.scss
@@ -32,6 +32,9 @@ $color-background-over-light: $fog;
 $color-text-light: $black;
 $color-border-light: #ccc;
 
+// Margins
+$margin-side: 0.935em;
+
 // Bootstrap vars from here (?)
 $primary: $cyan-text-dark;
