--- src/sections/Chat/PasswordLock/index.module.scss
@@ -1,45 +1,0 @@
-@import 'assets/scss/variables';
-
-$lockedColor: fade-out(#fff, 0.7);
-
-.passwordLock {
-  padding: 0.5em 0;
-}
-
-.unlocked {
-  color: $secondary !important;
-  animation: pulse linear 1s infinite;
-
-  &:hover {
-    color: scale-color($secondary, $lightness: -10%) !important;
-  }
-}
-
-.locked {
-  color: $lockedColor !important;
-
-  &:hover {
-    color: fade-in($lockedColor, 0.2) !important;
-  }
-}
-
-@keyframes pulse {
-  0% {
-    transform: scale(1);
-  }
-  50% {
-    transform: scale(1.1);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-
-.footer {
-  position: relative;
-
-  .left {
-    position: absolute;
-    left: 1em;
-  }
-}
