+++ src/sections/Chat/Session/PasswordModal/index.module.scss
@@ -1,0 +1,8 @@
+.footer {
+  position: relative;
+
+  .left {
+    position: absolute;
+    left: 1em;
+  }
+}
