+++ src/sections/Chat/index.tsx
@@ -7,23 +7,23 @@ import { useGlobalState } from 'store/state'
 import Loader from 'components/Loader'
 
 import styles from './index.module.scss'
+import Header from './Header'
 import Messages from './Messages'
 import TextBox from './TextBox'
 import OnlineUsers from './OnlineUsers'
-import PasswordLock from './PasswordLock'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useImmer } from 'use-immer'
 import { Message } from 'store/types'
 
 type State = {
-  isTextBoxFocused: boolean,
-  scrollingWhileFocused: boolean,
-  replyTo: Message | null,
+  isTextBoxFocused: boolean
+  scrollingWhileFocused: boolean
+  replyTo: Message | null
 }
 
 export default () => {
   const { state } = useGlobalState()
-  const [ localState, setState ] = useImmer<State>({
+  const [localState, setState] = useImmer<State>({
     isTextBoxFocused: false,
     scrollingWhileFocused: false,
     replyTo: null,
@@ -31,29 +31,36 @@ export default () => {
 
   let chatArea = <Loader />
 
-  const onMessageClick = useCallback((messageTimestamp: number) => {
-    setState(draft => {
-      const messageReply = state.messages.find(eachMessage => eachMessage.timestamp.getTime() === messageTimestamp) || null
-      console.log('replying to ', messageReply);
-      draft.replyTo = messageReply
-    })
-  }, [state.messages, setState])
+  const onMessageClick = useCallback(
+    (messageTimestamp: number) => {
+      setState((draft) => {
+        const messageReply =
+          state.messages.find(
+            (eachMessage) =>
+              eachMessage.timestamp.getTime() === messageTimestamp
+          ) || null
+        console.log('replying to ', messageReply)
+        draft.replyTo = messageReply
+      })
+    },
+    [state.messages, setState]
+  )
 
   const cancelReply = useCallback(() => {
-    setState(draft => {
+    setState((draft) => {
       draft.replyTo = null
     })
   }, [setState])
 
   // On window scroll
   const setHeaderPosition = useCallback(() => {
-    setState(draft => {
+    setState((draft) => {
       if (draft.isTextBoxFocused) {
         draft.scrollingWhileFocused = true
       }
     })
     // eslint-disable-next-line
-  }, [setState]);
+  }, [setState])
 
   useEffect(() => {
     window.addEventListener('scroll', setHeaderPosition, true)
@@ -64,14 +71,14 @@ export default () => {
   }, [setHeaderPosition])
 
   const handleTextBoxFocus = () => {
-    setState(draft => {
+    setState((draft) => {
       draft.isTextBoxFocused = true
     })
     setHeaderPosition()
   }
 
   const handleTextBoxBlur = () => {
-    setState(draft => {
+    setState((draft) => {
       draft.isTextBoxFocused = false
       draft.scrollingWhileFocused = false
     })
@@ -83,18 +90,22 @@ export default () => {
   } as React.CSSProperties
 
   if (localState.scrollingWhileFocused) {
-    headerStyle.position = 'absolute';
+    headerStyle.position = 'absolute'
     headerStyle.top = window.pageYOffset + 'px'
   }
 
   if (state.socket.connected) {
     chatArea = (
-      <div className={classNames(styles.chat, { [styles.scrollingChat]: localState.scrollingWhileFocused })}>
+      <div
+        className={classNames(styles.chat, {
+          [styles.scrollingChat]: localState.scrollingWhileFocused,
+        })}
+      >
+        <Header />
         <div className={classNames(styles.header)} style={headerStyle}>
           <Container>
             <Row>
               <Col>
-                <PasswordLock />
                 {state.auth.user && state.auth.user.isAdmin && (
                   <Link
                     to="/admin"
@@ -112,14 +123,12 @@ export default () => {
           </Container>
         </div>
 
-        <Messages
-          onMessageClick={ onMessageClick }
-        />
+        <Messages onMessageClick={onMessageClick} />
         <TextBox
-          replyTo={ localState.replyTo }
+          replyTo={localState.replyTo}
           onFocus={handleTextBoxFocus}
           onBlur={handleTextBoxBlur}
-          onCancelReply={ cancelReply }
+          onCancelReply={cancelReply}
         />
       </div>
     )
